added validateonly and boil receive delete functionality

+++ src/BuildingBlocks/EFCore.Support/EFCore.SQL/Repository/AccountToAssortMasterRepository.cs
@@ -36,7 +36,7 @@ namespace EFCore.SQL.Repository
             }
         }
 
-        public async Task<bool> DeleteAccountToAssortAsync(string accountToAssortId, string accountToAssortChildId, string slipNo)
+        public async Task<bool> DeleteAccountToAssortAsync(string accountToAssortId, string accountToAssortChildId, string slipNo, bool isValidateOnly = false)
         {
             using (_databaseContext = new DatabaseContext())
             {
@@ -44,10 +44,13 @@ namespace EFCore.SQL.Repository
 
                 var AssortDetailsRec = getMasterRecord.AccountToAssortDetails.Where(w => w.SlipNo == slipNo).FirstOrDefault();
 
-                var checkInBoil = await _databaseContext.BoilProcessMaster.Where(w => w.SlipNo == slipNo && w.AccountToAssortDetailsId == AssortDetailsRec.Id).ToListAsync();    
+                var checkInBoil = await _databaseContext.BoilProcessMaster.Where(w => w.SlipNo == slipNo && w.AccountToAssortDetailsId == AssortDetailsRec.Id).ToListAsync();
 
-                if(getMasterRecord != null && checkInBoil.Count == 0)
+                if (getMasterRecord != null && checkInBoil.Count == 0)
                 {
+                    if (isValidateOnly)
+                        return true;
+                    
                     _databaseContext.AccountToAssortDetails.RemoveRange(getMasterRecord.AccountToAssortDetails);
                     await _databaseContext.SaveChangesAsync();
